+++ test/puget/printer_test.clj
@@ -239,6 +239,20 @@
   (testing "map collection separator"
     (with-options {:map-coll-separator :line, :width 10}
       (is (= "{:bar\n [:a :b]}" (pprint-str {:bar [:a :b]})))))
+  (testing "namespace maps"
+    (with-options {:namespace-maps true}
+      (is (= "{:b 3, :a/x 1, :a/y 2}" (pprint-str {:a/x 1, :a/y 2, :b 3}))
+          "any simple keys should not be namespaced")
+      (is (= "#:a {:x 1, :y 2}" (pprint-str {:a/x 1, :a/y 2}))
+          "all qualified common ns keys should be namespaced")
+      (is (= "{:a/x 1, :b/x 2}" (pprint-str {:a/x 1, :b/x 2}))
+          "insufficiently common ns should not be qualified")
+      (is (= "#:a {:x 1, :y 2, :b/x 3}" (pprint-str {:a/x 1, :a/y 2, :b/x 3}))
+          "common ns should be qualified even with other ns keys")
+      (is (= "#:a {x 1, y 2}" (pprint-str {'a/x 1, 'a/y 2}))
+          "also works with symbol keys")
+      (is (= "{\"a/x\" 1, :a/y 2}" (pprint-str {"a/x" 1, :a/y 2}))
+          "any non-ident keys should not be namespaced")))
   (testing "lazy seq limits"
     (with-options {:seq-limit 4}
       (is (= "(1 2 3)" (pprint-str (map inc [0 1 2]))))