+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+...
+
+## [1.1.0] - 2018-12-16
+
 ### Added
 - The pretty-printer supports a `:namespace-maps` option which will render maps
   using the new namespace map literal syntax introduced in Clojure 1.9.0 when
@@ -151,7 +155,8 @@ to fipp and Clojure 1.7.
 - Maps default to using commas between entries.
 - Gave characters a separate syntax element than strings.
 
-[Unreleased]: https://github.com/greglook/puget/compare/1.0.3...HEAD
+[Unreleased]: https://github.com/greglook/puget/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/greglook/puget/compare/1.0.3...1.1.0
 [1.0.3]: https://github.com/greglook/puget/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/greglook/puget/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/greglook/puget/compare/1.0.0...1.0.1